requirements.txt
# Core
feedparser==6.0.11
requests==2.32.3          # kept during agent.py transition; remove after agent.py is retired
beautifulsoup4==4.12.3
trafilatura==1.9.0
PyYAML==6.0.2
python-dateutil==2.9.0.post0
rapidfuzz==3.9.6

# Async HTTP (replaces requests in new modules)
httpx>=0.27.0

# Async SQLite
aiosqlite>=0.20.0

# Numeric (cosine similarity)
numpy>=1.26.0

# Optional (only if you enable OpenAI summarization / embeddings)
openai>=1.65.0

# Email
